// src/app/core/models/equipo.ts
export interface Equipo{
  idEquipo: number;
  numeroInterno: string;
  placaInventario: string;
  placa: string;
  movil: string;
  idTipoClaseEquipo: number;
  idTipoEquipo: number;
  picoYPlaca: string;
  idTipoOrigenEquipo: number;
  plazoDiasMantenimiento: number;
  horasMantenimiento: number;
  kilometrosMantenimiento: number;
  fechaUltimoMantenimiento: number;
  fechaSiguienteMantenimiento: number;
  idTipoEstadoEquipo: number;
  idTipoMarcaEquipo: number;
  linea: string;
  cilindraje: number;
  numeroMotor: string;
  numeroChasis: string;
  idTipoCombustible: number;
  modelo: number;
  color: string;
  idTipoArea: number;
  fechaInicio: number;
  fechaFin: number;
  toneladas: number;
  pasajeros: number;
  numeroSerial: string;
  referencia: string;
  descripcion: string;
  volumenM3: number;
  idTipoUnidadUso: number;
}